+++ 框架埋点方案集成/springMVC-AOP/CatCacheTransaction.java
@@ -1,106 +1,11 @@
-import com.dianping.cat.Cat;
-import com.dianping.cat.CatConstants;
-import org.apache.ibatis.executor.Executor;
-import org.apache.ibatis.mapping.BoundSql;
-import org.apache.ibatis.mapping.MappedStatement;
-import org.apache.ibatis.mapping.ParameterMapping;
-import org.apache.ibatis.plugin.Interceptor;
-import org.apache.ibatis.plugin.Intercepts;
-import org.apache.ibatis.plugin.Invocation;
-import org.apache.ibatis.plugin.Plugin;
-import org.apache.ibatis.plugin.Signature;
-import org.apache.ibatis.reflection.MetaObject;
-import org.apache.ibatis.session.Configuration;
-import org.apache.ibatis.session.ResultHandler;
-import org.apache.ibatis.session.RowBounds;
-import org.apache.ibatis.type.TypeHandlerRegistry;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
-import java.text.DateFormat;
-import java.util.Date;
-import java.util.List;
-import java.util.Locale;
-import java.util.Properties;
-
-@Intercepts({
-        @Signature(type = Executor.class, method = "update", args = { MappedStatement.class, Object.class }),
-        @Signature(type = Executor.class, method = "query", args = { MappedStatement.class, Object.class,
-                RowBounds.class, ResultHandler.class }) })
-public class CatMybatisInterceptor implements Interceptor {
-
-    private Properties properties;
-
-    public Object intercept(Invocation invocation) throws Throwable {
-        //begin cat transaction
-        String datasourceUrl = properties.getProperty("datasourceUrl");
-
-        MappedStatement mappedStatement = (MappedStatement) invocation.getArgs()[0];
-        String sqlId = mappedStatement.getId();
-
-        com.dianping.cat.message.Transaction t = Cat.newTransaction(CatConstants.TYPE_SQL, sqlId);
-        Cat.logEvent("SQL.Database",datasourceUrl);
-        Cat.logEvent("SQL.Method",invocation.getMethod().getName());
-        try {
-            Object returnValue = invocation.proceed();
-            t.setStatus(com.dianping.cat.message.Transaction.SUCCESS);
-            return returnValue;
-        } catch (Throwable e) {
-            Cat.logError(e);
-            t.setStatus(e);
-            throw e;
-        }finally{
-            t.complete();
-        }
-
-    }
-
-    private static String getParameterValue(Object obj) {
-        String value = null;
-        if (obj instanceof String) {
-            value = "'" + obj.toString() + "'";
-        } else if (obj instanceof Date) {
-            DateFormat formatter = DateFormat.getDateTimeInstance(DateFormat.DEFAULT, DateFormat.DEFAULT, Locale.CHINA);
-            value = "'" + formatter.format(new Date()) + "'";
-        } else {
-            if (obj != null) {
-                value = obj.toString();
-            } else {
-                value = "";
-            }
-        }
-        return value;
-    }
-
-    public static String showSql(Configuration configuration, BoundSql boundSql) {
-        Object parameterObject = boundSql.getParameterObject();
-        List<ParameterMapping> parameterMappings = boundSql.getParameterMappings();
-        String sql = boundSql.getSql().replaceAll("[\\s]+", " ");
-        if (parameterMappings.size() > 0 && parameterObject != null) {
-            TypeHandlerRegistry typeHandlerRegistry = configuration.getTypeHandlerRegistry();
-            if (typeHandlerRegistry.hasTypeHandler(parameterObject.getClass())) {
-                sql = sql.replaceFirst("\\?", getParameterValue(parameterObject));
-
-            } else {
-                MetaObject metaObject = configuration.newMetaObject(parameterObject);
-                for (ParameterMapping parameterMapping : parameterMappings) {
-                    String propertyName = parameterMapping.getProperty();
-                    if (metaObject.hasGetter(propertyName)) {
-                        Object obj = metaObject.getValue(propertyName);
-                        sql = sql.replaceFirst("\\?", getParameterValue(obj));
-                    } else if (boundSql.hasAdditionalParameter(propertyName)) {
-                        Object obj = boundSql.getAdditionalParameter(propertyName);
-                        sql = sql.replaceFirst("\\?", getParameterValue(obj));
-                    }
-                }
-            }
-        }
-        return sql;
-    }
-
-    public Object plugin(Object target) {
-        return Plugin.wrap(target, this);
-    }
-
-    public void setProperties(Properties properties0) {
-        this.properties = properties0;
-    }
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface CatCacheTransaction {
+    String name() default "";
+    String server() default "Default";
 }
